Add files via upload
A manual supporting Win32Forth.  Using Word, this documents a major part of Win32Forth. It was written as I used W332F and had to examine source code to find the fine points. This is a never-to-be-done project.  There are many noted points of incompletion or ignorance.  If you take this document into another direction, at least give me credit.  But I'd rather have your additions and corrections sent to bill@billragsdale.cc.  Don't send me a complete markup. Just the changes and additions.