Add files via upload
This manual is being written for the Version 6.xx.xx of Win32Forth.  It is a work in process as I dredge added information from the source code.  It is released with no restrictions. Please give me credit if you make a derivative work.  If you know of a similar publication, please let me know.    bill at  billragsdale dot  cc.